+++ DTCD-StyleSystem/src/stories/BaseInput.stories.js
@@ -78,6 +78,7 @@ const Template = args => {
     labelSlot,
     invalid,
     maxlength,
+    minlength,
   } = args;
 
   const input = document.createElement(NAME_COMPONENT);
@@ -98,6 +99,7 @@ const Template = args => {
   input.readonly = readonly;
   input.invalid = invalid;
   input.maxlength = maxlength;
+  input.minlength = minlength;
 
   input.innerHTML += labelSlot;
   input.innerHTML += args['icon-left'];
@@ -122,6 +124,7 @@ DefaultInput.args = {
   'icon-left': '',
   'icon-right': '',
   maxlength: '',
+  minlength: '',
 };
 
 export const InputWithIcons = Template.bind({});
@@ -140,4 +143,5 @@ InputWithIcons.args = {
   'icon-left': '<span slot="icon-left" class="FontIcon name_searchSmall size_lg"></span>',
   'icon-right': '<span slot="icon-right" class="FontIcon name_show size_lg"></span>',
   maxlength: '',
+  minlength: '',
 };